requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific Computing
numpy>=1.21.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Data Handling
pandas>=1.5.0

# Visualization (optional but useful for analysis)
matplotlib>=3.6.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.64.0

# XML parsing (usually included with Python, but listing for completeness)
# xml.etree.ElementTree is part of Python standard library

# Logging and monitoring
tensorboard>=2.10.0

# Optional: for advanced image augmentations
# albumentations>=1.3.0

# Optional: for CUDA support (uncomment if using GPU)
# torch-audio>=0.13.0

# Development and testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0